+++ src/ewatercycle/forcing/_default.py
@@ -4,8 +4,6 @@ from copy import copy
 from pathlib import Path
 from typing import Optional
 
-from esmvalcore.experimental import CFG
-from esmvalcore.experimental.config import Session
 from ruamel.yaml import YAML
 
 from ewatercycle.util import to_absolute_path
@@ -91,20 +89,3 @@ class DefaultForcing:
 
     def __eq__(self, other):
         return self.__dict__ == other.__dict__
-
-
-def _session(directory: Optional[str] = None) -> Optional[Session]:
-    """When directory is set return a ESMValTool session that will write recipe output to that directory."""
-    if directory is None:
-        return None
-
-    class TimeLessSession(Session):
-        def __init__(self, output_dir: Path):
-            super().__init__(CFG.copy())
-            self.output_dir = output_dir
-
-        @property
-        def session_dir(self):
-            return self.output_dir
-
-    return TimeLessSession(Path(directory).absolute())
