+++ src/ewatercycle/plugins/lisflood/forcing.py
@@ -6,6 +6,9 @@ from typing import Optional
 
 from esmvalcore.experimental import get_recipe
 
+from ewatercycle.forcing._default import DefaultForcing
+from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.forcing.utils import _session
 from ewatercycle.util import (
     data_files_from_recipe_output,
     fit_extents_to_grid,
@@ -15,8 +18,6 @@ from ewatercycle.util import (
     to_absolute_path,
 )
 
-from ...forcing._default import DefaultForcing, _session
-from ...forcing.datasets import DATASETS
 from .lisvap import create_lisvap_config, lisvap
 
 logger = logging.getLogger(__name__)
