+++ src/ewatercycle/plugins/pcrglobwb/forcing.py
@@ -5,8 +5,9 @@ from typing import Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ewatercycle.forcing._default import DefaultForcing, _session
+from ewatercycle.forcing._default import DefaultForcing
 from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.forcing.utils import _session
 from ewatercycle.util import (
     data_files_from_recipe_output,
     get_extents,
