+++ src/ewatercycle/plugins/wflow/forcing.py
@@ -3,8 +3,9 @@ from typing import Dict, Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ewatercycle.forcing._default import DefaultForcing, _session
+from ewatercycle.forcing._default import DefaultForcing
 from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.forcing.utils import _session
 from ewatercycle.util import get_extents, get_time, to_absolute_path
